Multigrid method implemented; debugging required
The multigrid method is slow - slower than the Gauss-Seidel method originally implemented.  It should be noted that there is no stop condition in the solver for convergence of velocity and pressure.  Perhaps it has converged early, but it the solver kept going for no reason except that nothing told it to stop.